Gestion des exceptions

+++ src/main/java/com/example/api_bank/controller/CompteController.java
@@ -4,6 +4,7 @@ import com.example.api_bank.model.Client;
 import com.example.api_bank.model.Compte;
 import com.example.api_bank.repo.CompteRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,20 +24,38 @@ public class CompteController {
     }
 
     @GetMapping(value="/{numCpt}")
-    public Compte getCompteById(@PathVariable String numCpt){
-        return compteRepo.findById(numCpt).get();
+    public ResponseEntity<?> getCompteById(@PathVariable String numCpt) {
+        Optional<Compte> optionalCompte = compteRepo.findById(numCpt);
+        if (optionalCompte.isPresent()) {
+            Compte compte = optionalCompte.get();
+            return ResponseEntity.ok(compte);
+        } else {
+            String message = "compte  " + numCpt + " non trouvé";
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(message);
+        }
     }
 
     @DeleteMapping(value ="/{numCpt}")
-    public String supprimerCompte(@PathVariable String numCpt){
-        Compte compteSuppr = compteRepo.findById(numCpt).get();
-        compteRepo.delete(compteSuppr);
-        return "Compte spprimé id : "+ numCpt;
+    public ResponseEntity<?> supprimerCompte(@PathVariable String numCpt){
+        Optional<Compte> optionalCompte = compteRepo.findById(numCpt);
+        if (optionalCompte.isPresent()) {
+            Compte compteSuppr = optionalCompte.get();
+            compteRepo.delete(compteSuppr);
+            return ResponseEntity.ok().body("Compte " + numCpt + " supprimé");
+        } else {
+            String message = "Le compte " + numCpt + " non trouvé";
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(message);
+        }
     }
     @PostMapping
-    public String enregistrerCompte(@RequestBody Compte compte){
-        compteRepo.save(compte);
-        return "Enregistré ...";
+    public ResponseEntity<?> enregistrerCompte(@RequestBody Compte compte) {
+        try {
+            compteRepo.save(compte);
+            return ResponseEntity.ok("Compte " + compte.getNumCompte() + " créé");
+        } catch (Exception e) {
+            String errorMessage = "Erreur lors de l'enregistrement du compte: " + e.getMessage();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorMessage);
+        }
     }
 
 //    @PutMapping(value ="/comptes/{id}/modifier")
@@ -48,11 +67,15 @@ public class CompteController {
 //    }
 
     @GetMapping("/{numCpt}/proprietaire")
-    public Client getProprietaireByCompteId(@PathVariable  String numCpt) {
-        Compte compte = compteRepo.findById(numCpt).get();//.orElseThrow(() -> new ResourceNotFoundException("Compte", "id", compteId));
-        return compte.getProprietaire();
+    public ResponseEntity<?> getProprietaireByCompteId(@PathVariable String numCpt) {
+        Optional<Compte> optionalCompte = compteRepo.findById(numCpt);
+        if (optionalCompte.isPresent()) {
+            Compte compte = optionalCompte.get();
+            return ResponseEntity.ok(compte.getProprietaire());
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Compte " + numCpt + " non trouvé");
+        }
     }
-
 
     @PutMapping("/{numCompte}/retrait/{montant}")
     public ResponseEntity<?> faireRetrait(@PathVariable String numCompte, @PathVariable double montant) {
@@ -61,18 +84,20 @@ public class CompteController {
             Compte compte = optionalCompte.get();
             double nouveauSolde = compte.getSolde() - montant;
             if (nouveauSolde < 0) {
-                return ResponseEntity.badRequest().body("Le solde du compte ne peut pas être négatif");
+                return ResponseEntity.badRequest().body("Le solde du compte n'est pas suffisant pour un retrait ");
             }
             compte.setSolde(nouveauSolde);
             compteRepo.save(compte);
              return ResponseEntity.ok("Retrait effectué avec succès");
         } else {
-            return ResponseEntity.notFound().build();
+            String message = "Erreur lors de l'opération, compte "+numCompte+" non trouvé, relisez bien la documentation.";
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(message);
         }
     }
 
     @PutMapping("/{numCompte}/depot/{montant}")
     public ResponseEntity<?> faireDepot(@PathVariable String numCompte, @PathVariable double montant) {
+        try {
         Optional<Compte> optionalCompte = compteRepo.findById(numCompte);
         if (optionalCompte.isPresent()) {
             Compte compte = optionalCompte.get();
@@ -81,7 +106,15 @@ public class CompteController {
             compteRepo.save(compte);
             return ResponseEntity.ok("Dépôt effectué avec succès");
         } else {
-            return ResponseEntity.notFound().build();
+            String message = "compte " + numCompte + " non trouvé ";
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(message);
+        }
+
+
+        }
+        catch (Exception e) {
+            String message = "Erreur lors de l'opération, compte "+numCompte+" non trouvé , relisez bien la documentation.";
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(message);
         }
     }
 
@@ -94,7 +127,7 @@ public class CompteController {
             Compte compteDest = optionalCompteDest.get();
             double nouveauSoldeSource = compteSource.getSolde() - montant;
             if (nouveauSoldeSource < 0) {
-                return ResponseEntity.badRequest().body("Le solde du compte source ne peut pas être négatif");
+                return ResponseEntity.badRequest().body("Le solde du compte n'est pas suffisant pour un virement ");
             }
             double nouveauSoldeDest = compteDest.getSolde() + montant;
             compteSource.setSolde(nouveauSoldeSource);
@@ -102,7 +135,8 @@ public class CompteController {
             compteRepo.saveAll(Arrays.asList(compteSource, compteDest));
             return ResponseEntity.ok("Virement effectué avec succès");
         } else {
-            return ResponseEntity.notFound().build();
+            String message = "Erreur lors de l'opération,l'un ou les 2 comptes n'existent pas, relisez bien la documentation.";
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(message);
         }
     }
